[chip-tool] Create ValueChecker.h and ConstraintsChecker.h from methods signature living into TestCommand.h
Use CharSpan instead of Span<const char> and add a few const prefix
