+++ examples/chip-tool/commands/tests/TestCommand.cpp
@@ -95,88 +95,6 @@ void TestCommand::ThrowSuccessResponse()
     Exit("Expecting failure response but got a success response");
 }
 
-bool TestCommand::CheckConstraintType(const char * itemName, const char * current, const char * expected)
-{
-    ChipLogError(chipTool, "Warning: %s type checking is not implemented yet. Expected type: '%s'", itemName, expected);
-    return true;
-}
-
-bool TestCommand::CheckConstraintFormat(const char * itemName, const char * current, const char * expected)
-{
-    ChipLogError(chipTool, "Warning: %s format checking is not implemented yet. Expected format: '%s'", itemName, expected);
-    return true;
-}
-
-bool TestCommand::CheckConstraintStartsWith(const char * itemName, const chip::Span<const char> current, const char * expected)
-{
-    std::string value(current.data(), current.size());
-    if (value.rfind(expected, 0) != 0)
-    {
-        Exit(std::string(itemName) + " (\"" + value + "\") does not starts with: \"" + std::string(expected) + "\"");
-        return false;
-    }
-
-    return true;
-}
-
-bool TestCommand::CheckConstraintEndsWith(const char * itemName, const chip::Span<const char> current, const char * expected)
-{
-    std::string value(current.data(), current.size());
-    if (value.find(expected, value.size() - strlen(expected)) == std::string::npos)
-    {
-        Exit(std::string(itemName) + " (\"" + value + "\") does not ends with: \"" + std::string(expected) + "\"");
-        return false;
-    }
-
-    return true;
-}
-
-bool TestCommand::CheckConstraintMinLength(const char * itemName, uint64_t current, uint64_t expected)
-{
-    if (current < expected)
-    {
-        Exit(std::string(itemName) + " length < minLength: " + std::to_string(current) + " < " + std::to_string(expected));
-        return false;
-    }
-
-    return true;
-}
-
-bool TestCommand::CheckConstraintMaxLength(const char * itemName, uint64_t current, uint64_t expected)
-{
-    if (current > expected)
-    {
-        Exit(std::string(itemName) + " length > minLength: " + std::to_string(current) + " > " + std::to_string(expected));
-        return false;
-    }
-
-    return true;
-}
-
-bool TestCommand::CheckValueAsString(const char * itemName, chip::ByteSpan current, chip::ByteSpan expected)
-{
-    if (!current.data_equal(expected))
-    {
-        Exit(std::string(itemName) + " value mismatch, expecting " +
-             std::string(chip::Uint8::to_const_char(expected.data()), expected.size()));
-        return false;
-    }
-
-    return true;
-}
-
-bool TestCommand::CheckValueAsString(const char * itemName, chip::CharSpan current, chip::CharSpan expected)
-{
-    if (!current.data_equal(expected))
-    {
-        Exit(std::string(itemName) + " value mismatch, expected '" + std::string(expected.data(), expected.size()) + "' but got '" +
-             std::string(current.data(), current.size()) + "'");
-        return false;
-    }
-
-    return true;
-}
-
 bool TestCommand::ShouldSkip(const char * expression)
 {
     // If there is no PICS configuration file, considers that nothing should be skipped.
